Add swapping held item using number keys to inventory UI
The number keys can be used to swap the held item with the item in the item slot corresponding to the pressed number key.
Also fix problem where number keys couldn't be used to swap an item in the inventory UI when that item's slot was clicked twice without the mouse cursor exiting that item slot.