Update how player sprites are set based on movement input
Now, if the magnitude of the player movement input's y component is greater than the magnitude of that input's x component, then the sprite set will be based only on that input's y component. If that condition is false, then the sprite set will be based only on that input's x component. This change fixes problems where, for example, the player sprite would be facing right when the player movement input was up and slightly to the right.