Fix updating bool in FishingUIController.cs
Previously, the PlayerController.ActionDisablingUIOpen bool was set to false in FishingUIController.cs when the "Catch Failed" animation started. The fishing UI is open while that animation is playing, so that bool should be set to false after the "Catch Failed" animation finishes.
Also remove unnecessary return statement from FishingUIController.cs.