Allow item placement while mouse cursor is off-screen
Previously, while the mouse cursor was off-screen, item placement was disabled, and the in-game mouse cursor was visible at the edge of the screen with its default sprite. Now, after starting item placement, items can be placed at the edge of the screen by releasing the "Item Pickup and Placement" Action while the mouse cursor is off-screen.
Also:
- Rename MouseCanvasPositionHelper.cs to MousePositionHelper.cs and add the GetMousePositionClampedToScreen method to that script.
- Fix InventoryUITooltipController.cs using swing speed name instead of attack speed.