Update in-game mouse cursor for item pickup
Now, a red X is displayed on top of the item pickup arrow cursor while the player is attempting to pick up items with a full inventory.
Also remove unnecessary return statement from ItemPickupAndPlacement.cs and fix the IsFull method in Inventory.cs not returning an updated value in event handlers for the OnItemAdded event in that script.