Add full inventory check to CampfireItemInteractable.cs
Now, the inventory full UI is shown when a campfire recipe result item cannot fit in the player's inventory. Previously, the campfire recipe input item would be removed from the player's inventory and the campfire recipe result item would not be added to the player's inventory in that situation.
Also:
- Fix multiple campfire recipes potentially being applied to the player's inventory during campfire interaction.
- Update code for campfire recipe result items.