Fix in-game mouse cursor for item pickup
Previously, a red X was displayed on top of the item pickup arrow cursor while the player was attempting to pick up items with no empty item slots in their inventory. Now, a red X is displayed on top of the item pickup arrow cursor while that cursor is hovering over an item that cannot fit in the player's inventory.