Update dropping enemy loot
Now, an enemy dropping a loot item stack that cannot fully fit in the player's inventory results in a partial item stack that can fit in that inventory being added to that inventory and a remaining item stack being dropped in the world. Previously, the full loot item stack would be dropped in the world in that case.