Update campfire item interaction
Previously, the first item in the player's inventory that was an input item for a campfire recipe was cooked when the player interacted with a campfire. Now, one item in the selected item stack is cooked when the player interacts with a campfire if that item is an input item for a campfire recipe.